refactor: MainActivity feature.main 패키지로 이동
