+++ Near/app/src/main/java/com/alarmy/near/feature/main/MainActivity.kt
@@ -1,0 +1,21 @@
+package com.alarmy.near.feature.main
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.activity.enableEdgeToEdge
+import com.alarmy.near.ui.theme.NearTheme
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        enableEdgeToEdge()
+        setContent {
+            NearTheme {
+                NearApp()
+            }
+        }
+    }
+}
